software/neo/lumio/pyproject.toml
[project]
name = "lumio"
version = "2.0.0"
description = "Lumio V2 - Intelligent Educational Robot Platform"
requires-python = ">=3.11"
dependencies = [
    "RPi.GPIO>=0.7.1",
    "spidev>=3.7",
    "smbus2>=0.5.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "flask>=3.1.0",
    "flask-socketio>=5.5.0",
    "eventlet>=0.40.0",
    "pyserial>=3.5",
    "vosk>=0.3.45",
    "piper-tts>=1.2.0",
]

[project.scripts]
lumio = "lumio.main:main"

[tool.uv]
dev-dependencies = []
